Move log message helpers to LogUtils
This class will also contain other log helpers.
