+++ server/src/main/java/com/genymobile/scrcpy/CodecUtils.java
@@ -44,34 +44,6 @@ public final class CodecUtils {
         }
     }
 
-    public static String buildVideoEncoderListMessage() {
-        StringBuilder builder = new StringBuilder("List of video encoders:");
-        List<CodecUtils.DeviceEncoder> videoEncoders = CodecUtils.listVideoEncoders();
-        if (videoEncoders.isEmpty()) {
-            builder.append("\n    (none)");
-        } else {
-            for (CodecUtils.DeviceEncoder encoder : videoEncoders) {
-                builder.append("\n    --video-codec=").append(encoder.getCodec().getName());
-                builder.append(" --video-encoder='").append(encoder.getInfo().getName()).append("'");
-            }
-        }
-        return builder.toString();
-    }
-
-    public static String buildAudioEncoderListMessage() {
-        StringBuilder builder = new StringBuilder("List of audio encoders:");
-        List<CodecUtils.DeviceEncoder> audioEncoders = CodecUtils.listAudioEncoders();
-        if (audioEncoders.isEmpty()) {
-            builder.append("\n    (none)");
-        } else {
-            for (CodecUtils.DeviceEncoder encoder : audioEncoders) {
-                builder.append("\n    --audio-codec=").append(encoder.getCodec().getName());
-                builder.append(" --audio-encoder='").append(encoder.getInfo().getName()).append("'");
-            }
-        }
-        return builder.toString();
-    }
-
     private static MediaCodecInfo[] getEncoders(MediaCodecList codecs, String mimeType) {
         List<MediaCodecInfo> result = new ArrayList<>();
         for (MediaCodecInfo codecInfo : codecs.getCodecInfos()) {
